refactor: move slot-styles-mixin to component-base
