+++ packages/component-base/src/slot-styles-mixin.d.ts
